requirements.txt
Flask==1.1.2
plotly==4.14.3
streamlit==0.74.1
pandas==1.2.0
pillow==8.1.0